feat: add watchdog to helper
